+++ src/main/java/com/truelayer/java/paymentsproviders/entities/VrpCommercialCapabilities.java
@@ -1,7 +1,9 @@
 package com.truelayer.java.paymentsproviders.entities;
 
 import com.truelayer.java.entities.ProviderAvailability;
+import com.truelayer.java.entities.UseCase;
 import com.truelayer.java.payments.entities.ReleaseChannel;
+import java.util.List;
 import lombok.Value;
 
 @Value
@@ -9,4 +11,6 @@ public class VrpCommercialCapabilities {
     ReleaseChannel releaseChannel;
 
     ProviderAvailability availability;
+
+    List<UseCase> useCases;
 }