+++ src/main/java/com/truelayer/java/paymentsproviders/entities/searchproviders/VrpCommercialCapabilities.java
@@ -1,6 +1,10 @@
 package com.truelayer.java.paymentsproviders.entities.searchproviders;
 
+import com.truelayer.java.entities.UseCase;
+import java.util.List;
 import lombok.Value;
 
 @Value
-public class VrpCommercialCapabilities {}
+public class VrpCommercialCapabilities {
+    List<UseCase> useCases;
+}